add a spatial localizer base for asa and change marker spatial localizers to be reusable for different marker detectors
