+++ Assets/MixedRealityToolkit.Extensions/SpectatorView/Scripts/SpatialLocalizerBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 52f13462ab49cdb4b99dfc756cebd94a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
